Double download not possible anymore, and Searchbar now checks if its youtube link before commiting to backend.

+++ backend/harmonyhootenanny/modules/youtubedownloader.py
@@ -11,26 +11,34 @@ class YoutubeDownloader:
             yt = YouTube(youtube_link)
             stream = yt.streams.filter(only_audio=True, file_extension='mp4').first()
             if stream:
-                # Creates filename and downloads the video
-                mp3_path = f"{self.output_path}/{yt.title}.mp3"
-                stream.download(output_path=self.output_path, filename=f"{yt.title}.mp4")
-                # Renames the file, so it is in mp3 format
-                os.rename(f"{self.output_path}/{yt.title}.mp4", mp3_path)
-                
-                # Adds the information into the database
                 title = yt.title
                 artist = yt.author
                 duration = yt.length
                 file_name = f"{yt.title}.mp3"
-                
-                # Insert song into the database
-                add_song_to_db(title, artist, duration, file_name)
 
-                # Fetch the song ID TODO (assuming the song ID is returned by add_song_to_db)
+                # Check if the song already exists in the database
                 with get_db_connection() as conn:
                     cursor = conn.cursor()
-                    cursor.execute("SELECT id FROM songs WHERE title = ? AND artist = ? AND duration = ? AND src = ?", (title, artist, duration, file_name))
-                    song_id = cursor.fetchone()[0]
+                    cursor.execute("SELECT id FROM songs WHERE title = ?", (title,))
+                    existing_song = cursor.fetchone()
+
+                if existing_song:
+                    # Song already exists, do not download again
+                    song_id = existing_song[0]
+                else:
+                    # Song does not exist, download it
+                    mp3_path = f"{self.output_path}/{yt.title}.mp3"
+                    stream.download(output_path=self.output_path, filename=f"{yt.title}.mp4")
+                    os.rename(f"{self.output_path}/{yt.title}.mp4", mp3_path)
+
+                    # Insert song into the database
+                    add_song_to_db(title, artist, duration, file_name)
+
+                    # Fetch the song ID
+                    with get_db_connection() as conn:
+                        cursor = conn.cursor()
+                        cursor.execute("SELECT id FROM songs WHERE title = ? AND artist = ? AND duration = ? AND src = ?", (title, artist, duration, file_name))
+                        song_id = cursor.fetchone()[0]
 
                 # Add song to the queue
                 add_song_to_queue(song_id, roomId, userId)
